Lots of stuff to do with named items
I wanted a better way of working with variables in expressions.

This adds a "named items" screen, which shows all named items on the
stack and lets you edit the ones that are numbers. Ops can be named, and
their values are shown.

There's also now a keyboard shortcut @ to search for and add a named
item.

This also includes a few other changes.

This fixes a bug in the serialisation for very big numbers - it now uses
the long string version of numbers instead of the "nice number"
scientific notation.