trying to update to Vue 3
When I have a unary operation that works on another operation, I get a "Maximum recursion depth exceeded" error.

It's something to do with the fact that the CReal type is mutable, but I can't work out how to avoid it. I thought that making copies of the CReal objects used before calculating digits to display would do it, but apparently not.

The problem starts in the stack-item component: removing the <item-op> element stops the recursion error. Using a fixed CReal value as the :value attribute also stops the error.